+++ src/translation/en-US/index.ts
@@ -3,7 +3,6 @@ import {KitCollapseLocale} from './components/DataDisplay/Collapse';
 import {KitEmptyLocale} from './components/DataDisplay/Empty';
 import {KitErrorLocale} from '@translation/en-US/components/DataDisplay/Error';
 import {KitImageLocale} from './components/DataDisplay/Image';
-import {KitItemCardLocale} from './components/DataDisplay/ItemCard';
 import {KitItemListLocale} from './components/DataDisplay/ItemList';
 
 export const enUS: IKitLocale = {
@@ -12,6 +11,5 @@ export const enUS: IKitLocale = {
     Empty: KitEmptyLocale,
     Error: KitErrorLocale,
     Image: KitImageLocale,
-    ItemList: KitItemListLocale,
-    ItemCard: KitItemCardLocale
+    ItemList: KitItemListLocale
 };