+++ src/translation/types/index.d.ts
@@ -1,6 +1,5 @@
 import {Locale} from 'antd/es/locale';
 import {IKitItemListLocale} from './components/DataDisplay/ItemList';
-import {IKitItemCardLocale} from './components/DataDisplay/ItemCard';
 import {IKitImageLocale} from './components/DataDisplay/Image';
 import {IKitCollapseLocale} from './components/DataDisplay/Collapse';
 import {IKitEmptyLocale} from './components/DataDisplay/Empty';
@@ -15,5 +14,4 @@ export interface IKitLocale extends Locale {
     Error?: IKitErrorLocale;
     Image?: IKitImageLocale;
     ItemList?: IKitItemListLocale;
-    ItemCard?: IKitItemCardLocale;
 }