+++ stories/index.tsx
@@ -32,7 +32,6 @@ import {EditorTemplate as Empty} from './02-Design System/05-DataDisplay/Empty/T
 import {EditorTemplate as Error} from './02-Design System/05-DataDisplay/Error/Template';
 import {EditorTemplate as IdCard} from './02-Design System/05-DataDisplay/IdCard/Template';
 import {EditorTemplate as Image} from './02-Design System/05-DataDisplay/Image/Template';
-import {EditorTemplate as ItemCard} from './02-Design System/05-DataDisplay/ItemCard/Template';
 import {EditorTemplate as ItemList} from './02-Design System/05-DataDisplay/ItemList/Template';
 import {EditorTemplate as Table} from './02-Design System/05-DataDisplay/Table/Template';
 import {EditorTemplate as Tabs} from './02-Design System/05-DataDisplay/Tabs/Template';
@@ -89,7 +88,6 @@ export default {
         Error,
         IdCard,
         Image,
-        ItemCard,
         ItemList,
         Tabs,
         Table,